requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3
django-allauth>=0.60
social-auth-app-django>=5.4
dj-rest-auth>=4.0
djoser>=2.2
djangorestframework-simplejwt>=5.3
Pillow>=10.3  # अगर image upload या processing कर रहे हो


# asgiref==3.9.1
# certifi==2025.7.14
# cffi==1.17.1
# charset-normalizer==3.4.2
# cryptography==45.0.5
# defusedxml==0.7.1
# dj-rest-auth==7.0.1
# Django==5.2.4
# django-allauth==65.10.0
# django-cors-headers==4.7.0
# djangorestframework==3.16.0
# djangorestframework_simplejwt==5.5.0
# djoser==2.3.3
# idna==3.10
# oauthlib==3.3.1
# pycparser==2.22
# PyJWT==2.9.0
# python3-openid==3.2.0
# requests==2.32.4
# requests-oauthlib==2.0.0
# social-auth-app-django==5.5.1
# social-auth-core==4.7.0
# sqlparse==0.5.3
# tzdata==2025.2
# urllib3==2.5.0
